Don't report on pages that only live in content-store
travel-advice-publisher email_signup pages are new, and therefore not
in the contentapi. Therefore we shouldn't mistakenly report problems,
because it's fine for things not to be in the contentapi in the new
world. The taggings for these pages doesn't have to be in sync because
contentapi won't be used to display the tags.

https://www.gov.uk/api/content/foreign-travel-advice/portugal/email-sign
up